Refactor backend to clean architecture
