+++ backend/interfaces/api/auth.py
@@ -2,8 +2,8 @@
 
 from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.orm import Session
-from db.database import SessionLocal
-from db.models import User
+from infrastructure.database import SessionLocal
+from domain.models import User
 from schemas.auth import RegisterRequest, LoginRequest
 from utils.hash import hash_password, verify_password
