+++ backend/interfaces/api/stock.py
