+++ backend/usecases/strategy_executor/gain_ratio.py
@@ -1,5 +1,5 @@
 from typing import Dict, Any
-from services import kis_api
+from infrastructure import kis_api
 
 def execute_condition_gain_ratio(strategy: Dict[str, Any]) -> Dict[str, Any]:
     """
