+++ backend/usecases/strategy_executor/limit_buy.py
@@ -1,5 +1,5 @@
 from typing import Dict, Any
-from services import kis_api
+from infrastructure import kis_api
 import json
 
 def extract_float(value: Any) -> float:
